% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\encoding{UTF-8}
\name{append_vecOrList}
\alias{append_vecOrList}
\title{Appending two vector or list type objects.}
\usage{
append_vecOrList(x, item)
}
\arguments{
\item{x}{A vector or list type object.}

\item{item}{A vector or list type object appended to x.}
}
\description{
Appending two vector or list type objects.
}
\seealso{
Examples of this function can be found at \url{https://multivariate-statistics.com/2021/01/29/r-programming-vector-list/}
}
