% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\encoding{UTF-8}
\name{colBind}
\alias{colBind}
\title{Merging two objects (data.frame, vector) in a horizontal direction without adjusting the each number of rows or columns.}
\usage{
colBind(x, y, rowNames = TRUE, colNames = TRUE, sep = TRUE)
}
\arguments{
\item{x}{Data.frame type object or vector type object you want to merge with y.}

\item{y}{Data.frame type object or vector type object merged with x.}

\item{rowNames}{wheter to include rownames of x and y.}

\item{colNames}{wheter to include colnames of x and y.}

\item{sep}{Whether to separate x and y with a empty column.

#' @seealso Examples of this function can be found at \url{https://multivariate-statistics.com/2021/02/13/r-programming-rbind-cbind/}}
}
\description{
Merging two objects (data.frame, vector) in a horizontal direction without adjusting the each number of rows or columns.
}
