% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\encoding{UTF-8}
\name{colorCells_xlsx}
\alias{colorCells_xlsx}
\title{Coloring the columns specifying rows conditions}
\usage{
colorCells_xlsx(
  dataName,
  fileName,
  sheetName,
  coloredCols,
  coloredCondition = NULL,
  cellColor,
  fontSize = 11,
  fontName = "Yu Gothic",
  fontColor = "#000000"
)
}
\arguments{
\item{dataName}{The name of a csv-file you want to edit with coloring.}

\item{fileName}{The name of a Excel-file you want to save as.}

\item{sheetName}{The name of a EXCEL-sheet you want to color.}

\item{coloredCols}{The Columns colored with the specified color.}

\item{coloredCondition}{The condition for rows. This argument must be a logical vector in which TRUE components indicate row indices colored and neither FALSE components dose.}

\item{cellColor}{The color you want to color with}

\item{fontSize}{Font-size.}

\item{fontName}{Font-name.}

\item{fontColor}{The color of fonts.}
}
\description{
Coloring the columns specifying rows conditions
}
