% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\encoding{UTF-8}
\name{csvToxlsx}
\alias{csvToxlsx}
\title{Converting a csv file to a xlsx file}
\usage{
csvToxlsx(
  fileName_csv,
  fileName_xlsx,
  fontSize = 11,
  fontName = "Yu Gothic",
  fontColor = "#000000",
  fileEncoding = fileEncoding
)
}
\arguments{
\item{fileName_csv}{The name of a csv-file.}

\item{fileName_xlsx}{The name of a Excel-file.}

\item{fontSize}{Font-size.}

\item{fontName}{Font-name.}

\item{fontColor}{The color of fonts.}

\item{fileEncoding}{File-encoding.}
}
\description{
Converting a csv file to a xlsx file
}
