% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\encoding{UTF-8}
\name{dataCleanser}
\alias{dataCleanser}
\title{Cleansing the dataset on a csv-file to change its form to more arranged one to handle.}
\usage{
dataCleanser(
  dataName,
  append = FALSE,
  numOrFac = 10,
  leastNumOfDate = 10,
  fileEncoding = "CP932"
)
}
\arguments{
\item{dataName}{The file-name of a csv file that will be cleansed.}

\item{append}{Allows you to append the new datas generated from dataCleansingForm__.xlsx.}

\item{numOrFac}{The criteria for classifying whether the column data is numeric or factor. If the number of levels are greater than the ratio (nrow(data)/numOrFac), then it will be assiged to numeric group.}

\item{leastNumOfDate}{The criteria for classifying whether the column data is Date of numeric. if the data contains the dateFormat you have chosen and the number of data containing such formats is greater than this value, leastNumOfDate, then the data will be assigned to Date group.}

\item{fileEncoding}{File-encoding}
}
\description{
Cleansing the dataset on a csv-file to change its form to more arranged one to handle.
}
