% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\docType{class}
\encoding{UTF-8}
\name{dataCleansing-class}
\alias{dataCleansing-class}
\title{The class dataCleansing have several methods to operate the function dataCleanser.}
\description{
The class dataCleansing have several methods to operate the function dataCleanser.
}
\section{Fields}{

\describe{
\item{\code{cleansingForm}}{A form for data-cleansing stored into each types and each colnames.}

\item{\code{dataset}}{A dataset you want to organise.}
}}


