% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\encoding{UTF-8}
\name{excelHeadColor}
\alias{excelHeadColor}
\title{Coloring headers of tables on a EXCEL sheet}
\usage{
excelHeadColor(
  dataName,
  fileName,
  header,
  headerColor = "#92D050",
  fontSize = 11,
  fontName = "Yu Gothic",
  fontColor = "#000000",
  adj = TRUE,
  fileEncoding = "CP932"
)
}
\arguments{
\item{dataName}{The name of a csv-file you want to edit with coloring.}

\item{fileName}{The name of a Excel-file you want to save as.}

\item{header}{The character object included in each headers of tables.}

\item{headerColor}{The fore-ground-color of the headers of a data-frame.}

\item{fontSize}{Font-size.}

\item{fontName}{Font-name.}

\item{fontColor}{The color of fonts.}

\item{adj}{Allows you yo adjust shifted statistical test tables.}

\item{fileEncoding}{File-encoding.}
}
\description{
Coloring headers of tables on a EXCEL sheet
}
