% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\encoding{UTF-8}
\name{getCount}
\alias{getCount}
\title{Getting the number of a component in a vector or a list.}
\usage{
getCount(x, item)
}
\arguments{
\item{x}{A vector or list type object.}

\item{item}{A component in x.}
}
\description{
Getting the number of a component in a vector or a list.
}
\seealso{
Examples of this function can be found at \url{https://multivariate-statistics.com/2021/01/29/r-programming-vector-list/}
}
