% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\encoding{UTF-8}
\name{getIndex}
\alias{getIndex}
\title{Getting the indices of some item in a vector or a list.}
\usage{
getIndex(x, item)
}
\arguments{
\item{x}{A vector or list type object.}

\item{item}{The item in x indexed.}
}
\description{
Getting the indices of some item in a vector or a list.
}
\seealso{
Examples of this function can be found at \url{https://multivariate-statistics.com/2021/01/29/r-programming-vector-list/}
}
