% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\encoding{UTF-8}
\name{insert_vecOrList}
\alias{insert_vecOrList}
\title{Inserting an item into a vector or list type object.}
\usage{
insert_vecOrList(x, i, item)
}
\arguments{
\item{x}{A vector or list type object.}

\item{i}{The index which an item inserted at.}

\item{item}{A vector or list type object merged with x.}
}
\description{
Inserting an item into a vector or list type object.
}
