% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\encoding{UTF-8}
\name{isEqualData}
\alias{isEqualData}
\title{Compare two data in txt, csv or xlsx files.}
\usage{
isEqualData(
  fileName1,
  fileName2,
  fileEncoding1 = "CP932",
  fileEncoding2 = "CP932"
)
}
\arguments{
\item{fileName1}{The name of file in which one of data you want to compare is saved.}

\item{fileName2}{The name of file in which another data you want to compare is saved.}

\item{fileEncoding1}{File-encoding for fileName1.}

\item{fileEncoding2}{File-encoding for fileName2.}
}
\description{
Compare two data in txt, csv or xlsx files.
}
