% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\docType{package}
\name{<UsagiSan>}
\alias{<UsagiSan>}
\title{UsagiSan: A package for cleansing dataset and outputting statistical test results with using EXCEL.}
\description{
The package UsagiSan provides you a lot of helps to reduce the time on data-clansing and editting the test results. this package contains four function:
excelColor, excelHeadColor, mkDirectries and dataCleanser
}
\section{excelColor}{

The function excelColor helps you to edit the test results with coloring the signigicant variables with specific color.
}

\section{excelHeadColor}{

The function excelHeadColor helps you to add colors on headers of any type of tables including summaty sheets and statistical test tables.
}

\section{mkDirectries}{

The function mkDirectries organizes files in tha working directory.
}

\section{dataCleanser}{

The function dataCleanser helps you to cleans a dataset. This function modifies a dataset  to the form much easier to handle in statistical analysis.
}

\seealso{
\code{\link{excelColor}}

\code{\link{excelHeadColor}}

\code{\link{mkDirectries}}

\code{\link{dataCleanser}}

My web site: \url{https://multivariate-statistics.com}
}
