% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\name{mergeRowAndColnamesWithData}
\alias{mergeRowAndColnamesWithData}
\title{Merging rownames and colnames with a data.frame type object. This function reshapes a dataframe to the form having rownames and colnames as components of a dataframe not as its rownames and colnames.}
\usage{
mergeRowAndColnamesWithData(x)
}
\arguments{
\item{x}{A vector or dataframe type object you want to merge its rownames and colnames with.}
}
\description{
Merging rownames and colnames with a data.frame type object. This function reshapes a dataframe to the form having rownames and colnames as components of a dataframe not as its rownames and colnames.
}
