% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\encoding{UTF-8}
\name{mkDirectories}
\alias{mkDirectories}
\title{Making directories to organize three kinds of datas: data-sets,  script-files and result-files}
\usage{
mkDirectories(
  parentDirName,
  dataDirName = "data",
  programmingDirName = "program",
  resultDirName = "result",
  updateTime = 1,
  arrange = TRUE
)
}
\arguments{
\item{parentDirName}{The name of a parent-directory containing organize datas-files, script-files and result-files.}

\item{dataDirName}{The name of a directory to organize data-files.}

\item{programmingDirName}{The name of a directory to organize script-files.}

\item{resultDirName}{The name of a directory to organize result-files.}

\item{updateTime}{The time used to divide data-filese into two directories, one is for datas and the other is for results.}

\item{arrange}{Allows you to organize data-files in the form of file extensions.}
}
\description{
Making directories to organize three kinds of datas: data-sets,  script-files and result-files
}
