% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\encoding{UTF-8}
\name{mkDirectries}
\alias{mkDirectries}
\title{Making directries to organize three kinds of datas: data-sets,  script-files and result-files}
\usage{
mkDirectries(
  parentDirectryName,
  dataDirectryName = "data",
  programmingDirectryName = "program",
  resultDirectryName = "result",
  updateTime = 1,
  arrange = TRUE
)
}
\arguments{
\item{parentDirectryName}{The name of a parent-directry containing organize datas-files, script-files and result-files}

\item{dataDirectryName}{The name of a directry to organize data-files}

\item{programmingDirectryName}{The name of a directry to organize script-files}

\item{resultDirectryName}{The name of a directry to organize result-files}

\item{updateTime}{The time used to divide data-filese into two directries, one is for datas and the other is for results}

\item{arrange}{Allows you to organize data-files in the form of file extensions}
}
\description{
Making directries to organize three kinds of datas: data-sets,  script-files and result-files
}
