% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\encoding{UTF-8}
\name{pop_vecOrList}
\alias{pop_vecOrList}
\title{Popping an item from a vector or list type object.}
\usage{
pop_vecOrList(x, i)
}
\arguments{
\item{x}{A vector or list type object.}

\item{i}{The index of a component in x popped.}
}
\description{
Popping an item from a vector or list type object.
}
\seealso{
Examples of this function can be found at \url{https://multivariate-statistics.com/2021/01/29/r-programming-vector-list/}
}
