% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\encoding{UTF-8}
\name{remove_vecOrList}
\alias{remove_vecOrList}
\title{Removing an item in a vector or list type object.}
\usage{
remove_vecOrList(x, item)
}
\arguments{
\item{x}{A vector or list type object.}

\item{item}{An item removed from x.}
}
\description{
Removing an item in a vector or list type object.
}
\seealso{
Examples of this function can be found at \url{https://multivariate-statistics.com/2021/01/29/r-programming-vector-list/}
}
