% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\docType{package}
\name{<UsagiSan>}
\alias{<UsagiSan>}
\title{UsagiSan: A package for cleansing dataset and outputting statistical test results with using EXCEL.}
\description{
The package UsagiSan provides you a lot of helps to reduce the time on data-clansing and editting the test results. this package contains four function:
excelColor, excelHeadColor, mkDirectories and dataCleanser
}
\section{excelColor}{

The function excelColor helps you to edit the test results with coloring the signigicant variables with specific color.
}

\section{excelHeadColor}{

The function excelHeadColor helps you to add colors on headers of any type of tables including summaty sheets and statistical test tables.
}

\section{colorCells_xlsx}{

The function colorCells_xlsx colors columns with specified condition for rows in a EXCEL-sheet. This enables to color any EXCEL-sheets overwriting previous workbook data. you can freely intuitively edit EXCEL-sheets.
}

\section{mkDirectories}{

The function mkDirectories organizes files in tha working directory.
}

\section{dataCleanser}{

The function dataCleanser helps you to cleans a dataset. This function modifies a dataset  to the form much easier to handle in statistical analysis.
}

\section{dataCleansing}{

The class dataCleansing have several methods to operate dataCleanser. This provides how to classify a vector object into numeric, factor or Date type object.
}

\section{mergeRowAndColnamesWithData}{

The function mergeRowAndColnamesWithData reshapes a dataframe so that it has own rownames and colnames as its components.
}

\section{rowBind}{

The function rowBind is more useful and flexible than rbind. You can merge two objects without adjusting the number of rows or columns.
}

\section{colBind}{

The function colBind is more useful and flexible than cbind. You can merge two objects without adjusting the number of rows or columns.
}

\section{getIndex}{

The function getIndex gives you the indices of an item in a vector or a list.
}

\section{getCount}{

The function getCount gives you the number of an item in a vector or a list.
}

\section{append_vecOrList}{

The function append_vecOrList appends two vector or list type objects.
}

\section{extend_vecOrList}{

The function getCount extend_vecOrList two vector or list type objects to make a extended vector or list.
}

\section{insert_vecOrList}{

The function insert_vecOrList inserts an item at the specified index in a vector or list type object.
}

\section{remove_vecOrList}{

The function remove_vecOrList removes an item from a vector or list type object.
}

\section{pop_vecOrList}{

The function pop_vecOrList pops an item from a vector or list type object.
}

\seealso{
\code{\link{excelColor}}

\code{\link{excelHeadColor}}

\code{\link{excelHeadColor}}

\code{\link{mkDirectories}}

\code{\link{dataCleanser}}

\code{\linkS4class{dataCleansing}}

\code{\link{mergeRowAndColnamesWithData}}

\code{\link{rowBind}}

\code{\link{colBind}}

\code{\link{getIndex}}

\code{\link{getCount}}

\code{\link{append_vecOrList}}

\code{\link{extend_vecOrList}}

\code{\link{insert_vecOrList}}

\code{\link{remove_vecOrList}}

\code{\link{pop_vecOrList}}

My web site: \url{https://multivariate-statistics.com}
}
