% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\docType{package}
\name{<UsagiSan>}
\alias{<UsagiSan>}
\title{UsagiSan: A package for cleansing dataset and outputting statistical test results with using EXCEL.}
\description{
The package UsagiSan provides you a lot of helps to reduce the time on data-clansing and editting the test results. this package contains four function:
excelColor, excelHeadColor, mkDirectories and dataCleanser
}
\section{excelColor}{

The function excelColor helps you to edit the test results with coloring the signigicant variables with specific color.
}

\section{excelHeadColor}{

The function excelHeadColor helps you to add colors on headers of any type of tables including summaty sheets and statistical test tables.
}

\section{mkDirectories}{

The function mkDirectories organizes files in tha working directory.
}

\section{dataCleanser}{

The function dataCleanser helps you to cleans a dataset. This function modifies a dataset  to the form much easier to handle in statistical analysis.
}

\section{dataCleansing}{

The class dataCleansing have several methods to operate dataCleanser. This provides how to classify a vector object into numeric, factor or Date type object.
}

\section{mergeRowAndColnamesWithData}{

The function mergeRowAndColnamesWithData reshapes a dataframe so that it has own rownames and colnames as its components.
}

\section{rowBind}{

The function rowBind is more useful and flexible than rbind. You can merge two objects without adjusting the number of rows or columns.
}

\section{colBind}{

The function colBind is more useful and flexible than cbind. You can merge two objects without adjusting the number of rows or columns.
}

\section{getIndex}{

The function getIndex gives you the indices of some item in a vector or a list.
}

\seealso{
\code{\link{excelColor}}

\code{\link{excelHeadColor}}

\code{\link{mkDirectories}}

\code{\link{dataCleanser}}

\code{\linkS4class{dataCleansing}}

\code{\link{mergeRowAndColnamesWithData}}

\code{\link{rowBind}}

\code{\link{colBind}}

\code{\link{getIndex}}

My web site: \url{https://multivariate-statistics.com}
}
