% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagiSan.R
\encoding{UTF-8}
\name{rowBind}
\alias{rowBind}
\title{Merging two objects (data.frame, vector) in a vertical direction without adjusting the each number of rows or columns.}
\usage{
rowBind(x, y, sep = TRUE)
}
\arguments{
\item{x}{Data.frame type object or vector type object you want to merge with y.}

\item{y}{Data.frame type object or vector type object merged with x}

\item{sep}{Whether you separate x and y with a empty column}
}
\description{
Merging two objects (data.frame, vector) in a vertical direction without adjusting the each number of rows or columns.
}
